Refactor rawlog-edit as public C++ class
