+++ libs/apps/src/rawlog-edit_enose.cpp
@@ -7,24 +7,29 @@
    | Released under BSD License. See: https://www.mrpt.org/License          |
    +------------------------------------------------------------------------+ */
 
-#include <mrpt/obs/CObservationOdometry.h>
+#include "apps-precomp.h"  // Precompiled headers
+//
+
+#include <mrpt/containers/stl_containers_utils.h>
+#include <mrpt/obs/CObservationGasSensors.h>
 
 #include "rawlog-edit-declarations.h"
 
 using namespace mrpt;
 using namespace mrpt::obs;
 using namespace mrpt::system;
-using namespace mrpt::rawlogtools;
+using namespace mrpt::apps;
 using namespace std;
 using namespace mrpt::io;
 
 // ======================================================================
-//		op_export_txt
+//		op_export_enose_txt
 // ======================================================================
-DECLARE_OP_FUNCTION(op_export_txt)
+DECLARE_OP_FUNCTION(op_export_enose_txt)
 {
 	// A class to do this operation:
-	class CRawlogProcessor_Export_TXT : public CRawlogProcessorOnEachObservation
+	class CRawlogProcessor_ExportENOSE_TXT
+		: public CRawlogProcessorOnEachObservation
 	{
 	   protected:
 		string m_inFile;
@@ -35,7 +40,8 @@ DECLARE_OP_FUNCTION(op_export_txt)
 	   public:
 		size_t m_entriesSaved;
 
-		CRawlogProcessor_Export_TXT(
+		// Default Constructor
+		CRawlogProcessor_ExportENOSE_TXT(
 			CFileGZInputStream& in_rawlog, TCLAP::CmdLine& cmdline,
 			bool Verbose)
 			: CRawlogProcessorOnEachObservation(in_rawlog, cmdline, Verbose),
@@ -47,32 +53,27 @@ DECLARE_OP_FUNCTION(op_export_txt)
 				extractFileDirectory(m_inFile) + extractFileName(m_inFile);
 		}
 
-		// return false on any error.
-		bool processOneObservation(CObservation::Ptr& obs) override
+		// Return false on any error.
+		bool processOneObservation(CObservation::Ptr& o) override
 		{
-			if (!obs->exportTxtSupported()) return true;
+			if (!IS_CLASS(*o, CObservationGasSensors)) return true;
 
-			ASSERTMSG_(
-				!obs->sensorLabel.empty(),
-				mrpt::format(
-					"Exporting as TXT requires all observations having "
-					"non-empty sensorLabels, but empty label found for "
-					"observation of type '%s'",
-					obs->GetRuntimeClass()->className));
+			const CObservationGasSensors* obs =
+				dynamic_cast<CObservationGasSensors*>(o.get());
 
 			auto it = lstFiles.find(obs->sensorLabel);
 
-			FILE* f_this = nullptr;
+			FILE* f_this;
 
 			if (it == lstFiles.end())  // A new file for this sensorlabel??
 			{
 				const std::string fileName = m_filPrefix + string("_") +
 					fileNameStripInvalidChars(obs->sensorLabel.empty()
-												  ? string("ODOMETRY")
+												  ? string("ENOSE")
 												  : obs->sensorLabel) +
 					string(".txt");
 
-				VERBOSE_COUT << "Writing TXT/CSV file: " << fileName << endl;
+				VERBOSE_COUT << "Writing e-nose TXT file: " << fileName << endl;
 
 				f_this = lstFiles[obs->sensorLabel] =
 					os::fopen(fileName.c_str(), "wt");
@@ -82,40 +83,68 @@ DECLARE_OP_FUNCTION(op_export_txt)
 						fileName.c_str());
 
 				// The first line is a description of the columns:
-				::fprintf(
-					f_this,
-					"%% "
-					"%16s "	 // TIMESTAMP
-					"%s\n",
-					"Time", obs->exportTxtHeader().c_str());
+				//------------------------------------------------
+				// Time:
+				::fprintf(f_this, "%% %13s ", "Time");
+
+				// For each E-nose (if more than one)
+				for (size_t j = 0; j < obs->m_readings.size(); j++)
+				{
+					// Temperature
+					::fprintf(f_this, " Temp%u ", static_cast<unsigned int>(j));
+
+					// For each sensor on the E-nose
+					for (size_t k = 0;
+						 k < obs->m_readings[j].readingsVoltage.size(); k++)
+						::fprintf(
+							f_this, " S%u_%u ", static_cast<unsigned int>(j),
+							static_cast<unsigned int>(k));
+				}
+				::fprintf(f_this, "\n");
 			}
 			else
 				f_this = it->second;
 
+			// For each entry in this sequence: Compute the timestamp and save
+			// all values:
 			ASSERT_(obs->timestamp != INVALID_TIMESTAMP);
-			double sampleTime = mrpt::Clock::toDouble(obs->timestamp);
+			TTimeStamp t = obs->timestamp;
+
+			double sampleTime = timestampTotime_t(t);
 
 			// Time:
-			::fprintf(
-				f_this,
-				"%16.6f "  // TIMESTAMP
-				"%s\n",
-				sampleTime, obs->exportTxtDataRow().c_str());
+			::fprintf(f_this, "%14.4f ", sampleTime);
+
+			// For each E-nose (if more than one)
+			for (const auto& m_reading : obs->m_readings)
+			{
+				// Temperature
+				float temp = 0.0;
+				if (m_reading.hasTemperature == true)
+					temp = m_reading.temperature;
+				::fprintf(f_this, "%3.4f ", temp);
+
+				// For each sensor on the E-nose
+				mrpt::containers::printSTLContainer(m_reading.readingsVoltage);
+			}
+
+			::fprintf(f_this, "\n");
 			m_entriesSaved++;
 			return true;  // All ok
 		}
 
 		// Destructor: close files and generate summary files:
-		~CRawlogProcessor_Export_TXT()
+		~CRawlogProcessor_ExportENOSE_TXT()
 		{
 			for (auto it = lstFiles.begin(); it != lstFiles.end(); ++it)
-				if (it->second) os::fclose(it->second);
+			{
+				os::fclose(it->second);
+			}
 
 			// Save the joint file:
 			// -------------------------
-			VERBOSE_COUT
-				<< "Number of different sensorLabels exported to TXT/CSV: "
-				<< lstFiles.size() << endl;
+			VERBOSE_COUT << "Number of different E-nose sensorLabels  : "
+						 << lstFiles.size() << endl;
 
 			lstFiles.clear();
 		}  // end of destructor
@@ -123,7 +152,7 @@ DECLARE_OP_FUNCTION(op_export_txt)
 
 	// Process
 	// ---------------------------------
-	CRawlogProcessor_Export_TXT proc(in_rawlog, cmdline, verbose);
+	CRawlogProcessor_ExportENOSE_TXT proc(in_rawlog, cmdline, verbose);
 	proc.doProcessRawlog();
 
 	// Dump statistics:
