+++ libs/apps/src/rawlog-edit_list-timestamps.cpp
@@ -7,81 +7,63 @@
    | Released under BSD License. See: https://www.mrpt.org/License          |
    +------------------------------------------------------------------------+ */
 
-#include <mrpt/obs/CObservation3DRangeScan.h>
-
+#include "apps-precomp.h"  // Precompiled headers
+//
 #include "rawlog-edit-declarations.h"
 
 using namespace mrpt;
 using namespace mrpt::obs;
 using namespace mrpt::system;
-using namespace mrpt::rawlogtools;
+using namespace mrpt::apps;
 using namespace std;
 using namespace mrpt::io;
 
 // ======================================================================
-//		op_generate_3d_pointclouds
+//		op_list_timestamps
 // ======================================================================
-DECLARE_OP_FUNCTION(op_generate_3d_pointclouds)
+DECLARE_OP_FUNCTION(op_list_timestamps)
 {
 	// A class to do this operation:
-	class CRawlogProcessor_Generate3DPointClouds
+	class CRawlogProcessor_ListTimestamps
 		: public CRawlogProcessorOnEachObservation
 	{
 	   protected:
-		TOutputRawlogCreator outrawlog;
+		string m_out_file;
+		std::ofstream m_out;
 
 	   public:
-		size_t entries_modified;
-
-		CRawlogProcessor_Generate3DPointClouds(
+		CRawlogProcessor_ListTimestamps(
 			CFileGZInputStream& in_rawlog, TCLAP::CmdLine& cmdline,
 			bool Verbose)
 			: CRawlogProcessorOnEachObservation(in_rawlog, cmdline, Verbose)
 		{
-			entries_modified = 0;
+			getArgValue<std::string>(cmdline, "text-file-output", m_out_file);
+			VERBOSE_COUT << "Writing list to: " << m_out_file << endl;
+
+			m_out.open(m_out_file.c_str());
+
+			if (!m_out.is_open())
+				throw std::runtime_error(
+					"list-timestamps: Cannot open output text file.");
 		}
 
 		bool processOneObservation(CObservation::Ptr& obs) override
 		{
-			if (IS_CLASS(*obs, CObservation3DRangeScan))
-			{
-				CObservation3DRangeScan::Ptr obs3D =
-					std::dynamic_pointer_cast<CObservation3DRangeScan>(obs);
-				if (obs3D->hasRangeImage)
-				{
-					obs3D->load();	// We must be sure that depth has been
-					// loaded, if stored separately.
-					obs3D->unprojectInto(*obs3D);
-					entries_modified++;
-				}
-			}
-
+			m_out << std::fixed
+				  << mrpt::system::timestampToDouble(obs->timestamp) << " "
+				  << obs->sensorLabel << " "
+				  << obs->GetRuntimeClass()->className << std::endl;
 			return true;
-		}
-
-		// This method can be reimplemented to save the modified object to an
-		// output stream.
-		void OnPostProcess(
-			mrpt::obs::CActionCollection::Ptr& actions,
-			mrpt::obs::CSensoryFrame::Ptr& SF,
-			mrpt::obs::CObservation::Ptr& obs) override
-		{
-			ASSERT_((actions && SF) || obs);
-			if (actions) (*outrawlog.out_rawlog) << actions << SF;
-			else
-				(*outrawlog.out_rawlog) << obs;
 		}
 	};
 
 	// Process
 	// ---------------------------------
-	CRawlogProcessor_Generate3DPointClouds proc(in_rawlog, cmdline, verbose);
+	CRawlogProcessor_ListTimestamps proc(in_rawlog, cmdline, verbose);
 	proc.doProcessRawlog();
 
 	// Dump statistics:
 	// ---------------------------------
 	VERBOSE_COUT << "Time to process file (sec)        : " << proc.m_timToParse
 				 << "\n";
-	VERBOSE_COUT << "Entries modified                  : "
-				 << proc.entries_modified << "\n";
 }
